+++ src/app/settingsdialog.h
@@ -5,6 +5,7 @@
 
 class QCheckBox;
 class QComboBox;
+class QLabel;
 class QLineEdit;
 
 class SettingsDialog : public QDialog
@@ -16,12 +17,18 @@ public:
 
 private slots:
     void chooseLogDirectory();
+    void copyCodexMcpCommand();
+    void copyClaudeMcpCommand();
+    void copyMcpJson();
     void save();
 
 private:
+    void copyMcpText(const QString &text, const QString &description);
+
     QComboBox *m_terminalTheme = nullptr;
     QCheckBox *m_sessionLogging = nullptr;
     QLineEdit *m_logDirectory = nullptr;
     QCheckBox *m_restoreWorkspace = nullptr;
     QCheckBox *m_reconnectSsh = nullptr;
+    QLabel *m_mcpCopyStatus = nullptr;
 };